Fixed page footer position(Y) issue

+++ pkg/pdf/pdf.go
@@ -184,10 +184,10 @@ func (s *PdfMaroto) AddPage() {
 	_, pageHeight := s.Pdf.GetPageSize()
 	_, top, _, bottom := s.Pdf.GetMargins()
 
-	totalOffsetY := int(s.offsetY + s.footerHeight)
-	maxOffsetPage := int(pageHeight - bottom - top)
+	totalOffsetY := (s.offsetY + s.footerHeight)
+	maxOffsetPage := (pageHeight - bottom - top)
 
-	s.Row(float64(maxOffsetPage-totalOffsetY), func() {
+	s.Row(maxOffsetPage-totalOffsetY, func() {
 		s.ColSpace(uint(consts.MaxGridSum))
 	})
 }
@@ -359,8 +359,8 @@ func (s *PdfMaroto) Row(height float64, closure func()) {
 	_, pageHeight := s.Pdf.GetPageSize()
 	_, top, _, bottom := s.Pdf.GetMargins()
 
-	totalOffsetY := int(s.offsetY + height + s.footerHeight)
-	maxOffsetPage := int(pageHeight - bottom - top)
+	totalOffsetY := (s.offsetY + height + s.footerHeight)
+	maxOffsetPage := (pageHeight - bottom - top)
 
 	// Note: The headerFooterContextActive is needed to avoid recursive
 	// calls without end, because footerClosure and headerClosure actually
@@ -634,10 +634,10 @@ func (s *PdfMaroto) drawLastFooter() {
 		_, top, _, bottom := s.Pdf.GetMargins()
 
 		if s.offsetY+s.footerHeight < pageHeight-bottom-top {
-			totalOffsetY := int(s.offsetY + s.footerHeight)
-			maxOffsetPage := int(pageHeight - bottom - top)
+			totalOffsetY := (s.offsetY + s.footerHeight)
+			maxOffsetPage := (pageHeight - bottom - top)
 
-			s.Row(float64(maxOffsetPage-totalOffsetY), func() {
+			s.Row(maxOffsetPage-totalOffsetY, func() {
 				s.ColSpace(12)
 			})
 
@@ -655,10 +655,10 @@ func (s *PdfMaroto) footer() {
 	_, pageHeight := s.Pdf.GetPageSize()
 	_, top, _, bottom := s.Pdf.GetMargins()
 
-	totalOffsetY := int(s.offsetY + s.footerHeight)
-	maxOffsetPage := int(pageHeight - bottom - top)
+	totalOffsetY := (s.offsetY + s.footerHeight)
+	maxOffsetPage := (pageHeight - bottom - top)
 
-	s.Row(float64(maxOffsetPage-totalOffsetY), func() {
+	s.Row(maxOffsetPage-totalOffsetY, func() {
 		s.ColSpace(uint(consts.MaxGridSum))
 	})
